[OG-74] Forgot about providers
