+++ polymorphia-frontend/app/(logged-in)/layout.tsx
@@ -2,7 +2,7 @@
 
 import { ReactNode } from "react";
 import Navigation from "@/components/navigation/Navigation";
-import { NavigationProvider } from "@/components/providers/NavigationContext";
+import { NavigationProvider } from "@/components/providers/navigation/NavigationContext";
 import Image from "next/image";
 import MainLayout from "@/components/main-layout/MainLayout";
 import { UserProvider } from "@/components/providers/user/UserContext";
