+++ polymorphia-frontend/app/(logged-in)/settings/page.tsx
@@ -2,7 +2,7 @@
 
 import ButtonWithBorder from "@/components/button/ButtonWithBorder";
 import { useContext, useEffect } from "react";
-import { NavigationContext } from "@/components/providers/NavigationContext";
+import { NavigationContext } from "@/components/providers/navigation/NavigationContext";
 import { useScaleShow } from "@/animations/General";
 import { useTitle } from "@/components/navigation/TitleContext";
 import toast from "react-hot-toast";
