+++ polymorphia-frontend/components/navigation/MenuSection.tsx
@@ -1,6 +1,6 @@
 import { useContext, useEffect, useRef, useState } from "react";
 import { ChevronDown, ChevronUp } from "lucide-react";
-import { NavigationContext } from "@/components/providers/NavigationContext";
+import { NavigationContext } from "@/components/providers/navigation/NavigationContext";
 import { MenuSectionProps } from "@/components/navigation/types";
 import { animateSubMenuSection } from "@/animations/Navigation";
 import Link from "next/link";
