+++ polymorphia-frontend/components/providers/modal/types.ts
@@ -1,0 +1,8 @@
+export interface ModalContextValue {
+  closeModal: () => void;
+}
+
+export interface ModalProviderProps {
+  children: React.ReactNode;
+  closeModal: () => void;
+}
